Prototype implementation for a general "big monster" api
This introduces the beginnings of a general api for monsters that can
occupy more than one cell, along with an initial type of big_monster,
line_monster. The implementation is via a class `big_monster` that sits
on top of regular `monster`s, tracking parts and controlling behavior in
various ways. big_monsters (for the moment) are always structured as a
`head` monster and a set of parts, and their details are serialized on
the head monster. The parts generally are not autonomous, may come and
go (similar to tentacles), and inherit various properties of the head
monster (though implementing this fully is pretty ambitious and no doubt
this first version misses various things). Right now big_monsters are
intended to have flexible sizes, always including the possibility of
occupying one cell (inflexible multi-cell monsters may eventually be
possible, but this is even harder within the constraints of crawl's
existing monster api).

Specific behavior patterns are implemented by subclassing big_monster.
This commit provides one such subclass, `line_monster`. line_monsters
have a snake-like behavior, where the parts trail along behind the head,
except that I have implemented things so that the head can move through
parts. They attack only from the head. Two monsters are currently
hard-coded to generate with this behavior, iguanas (max length 2) and
worms (max length 6); eventually this needs to be specifiable on the
monster info itself, but this isn't ready yet. I have a few more
behavior patterns in the works. (And one eventual goal is a cleaner
rewrite of tentacled monsters.)

Tiles rendering is not yet implemented in this commit, probably along
with all sorts of other stuff I've missed.