monstercrawl: initial patch
An experiment in player-monster symmetry.

This patch implements a core idea with many details still to be worked
out: play as a dcss monster. Your first instinct with this idea (as seen
in forks like gnollcrawl, oofcrawl) might be to simply implement
monsters as player species. However, this is unsustainable if your goal
is to make all monsters playable (this is my goal). Therefore, I have
come up with another way.  The basic technique is to merge monsters and
players. This patch replaces player species with a fancy union class
that also stores a monster species, and then modifies many places in the
code so that they can then apply to the player using monster functions.
Because monster species by itself doesn't provide all the information
you need about monster properties, we also keep around a full instance
of the monster type tacked on to the player as well -- basically
generalizing the technique from dith shadows.

This patch represents a lot of work on many special cases for monster
behavior, and it broadly should cover attacking, movement, and many
passive effects for most monsters. The biggest glaring hole is monster
abilities/spells (which at the time of committing this are waiting on a
merge with some updated code in crawl master). However, it still needs a
ton more testing even for the implemented things.

There are various choices I made here that could have been different;
the biggest one is probably to institute some kind of leveling system
for monster players. (The alternative would be to just play as the
monster with minimal growth, gnollcrawl-style.) I'm sure this still
needs some tweaking. Needless to say, even with this leveling system the
goal is not to have a balanced range of playable options, unless by
balanced you mean, ranging from completely, utterly, unplayable
(statues, eels, butterflies) to extremely OP even at xl 1 (cerebov,
oofs, shrikes, etc). However, I concluded that without some kind of
character progression beyond what is found in the dungeon, this fork
would be just a gimmick, *maybe* with it it will have replayability.