feat: implement Bai Suzhen transformation
This is a bit of a departure from monster Bai Suzhen. In her case, once
she hits half hp, the transformation to dragon form is permanent. This
seemed less interesting to me for the player, so I implemented it as a
hybrid of species change and dragon form. When player Bai Suzhen is in
dragon form, she temporarily becomes species MON_BAI_SUZHEN_DRAGON with
all the accompanying bells and whistles, and reverts when the form ends.
When player Bai Suzhen hits half hp, dragon form is automatically
triggered. This does *not* change skill aptitudes, basically because I
couldn't get that work right even with a fair amount of effort. Items
act melded as usual.

It is still possible to play as the dragon only, by entering "bai suzhen
dragon" as the monster description when starting play.

This commit also implements silver damage / Zin dislike for chaotic
monster players, which impacts Bai Suzhen but also many others.