feat: implement shapeshifters
I thought I might do this with the base_monster_instance infrastructure,
but it ended up being easier to just follow the monster implementation.
The shapeshifter monster identity is carried on two enchantment types,
and a player shapeshifter has their monster species fully changed on the
same schedule as monster shapeshifters. I'm not really sure how playable
this is with this timing, and possibly some spam could be reduced.
There's also other awkward things like skill changes that may need some
work, and the naming code for these is a bit all over the place. But,
the core works and saves/loads.

This ignores shapechanging messages during runrest delays by default by
moving it into an ignored channel; it's slightly hacky in that there is
no perfectly natural channel for this.