feat: don't print `Monster` in the save browser
Since it's hard to get a regular player species now, this is just adding
horizontal filler to the save browser for the most part. Regular species
that might be confusable would still be differentiated in that they will
show a job. (Could adapt the convention from other parts of the current
implementation of showing "Monstrous" for these.)

Also, improve the labeling of wizmode-generated characters that have
JOB_MONSTER. (Which should otherwise be impossible, this isn't a real
job.)