feat: spriggan riders v1
This commit implements spriggan riders via the lives machinery. A major
departure from the monster version is that, for now, these always
convert to a spriggan (as opposed to a coinflip between hornet and
spriggan). This might change, but it seems to me that switching to a
hornet will make the species relatively unplayable -- it'll make some
training paths useless, make items useless, etc. whereas a spriggan
rider already uses spriggan apts and so on. Another departure from
rebirth species is that bring_to_safety is skipped, given the flavor of
falling from the dead hornet. Arguably Bennu should get this also.

This commit also includes a big refactor of monster species name
construction, which was getting very clunky. There are still aspects of
the bolted-on initial implementation (monster "job" is still used to
render the monster species, and species name is still just "Monster")
that could use further refinement, but this at least centralizes most of
the custom code.

Some missing flavor: the hornet corpse should get named