refactor: separate monster apt initialization from internals
This moves it to its own file. I'm planning on moving other
initialization stuff here as well, but the apts in particular were
enough of an endeavor that it seems better to put this in its own
commit. I cleaned up this code and fixed a number of bugs about how
player genus species were handled relative to the original intent,
though this commit doesn't make any major changes to initialization.