feat: various monsters' self-destructive tendencies
This is pretty silly, but this level of attention to detail is what this
fork is about. The only one of these that is conceivably viable is orb
of destruction, which will not die unless you move it into a monster. (I
didn't implement wall collision, which would make this species even more
of an exercise in not mistyping.)

Covers:
* OoD destruction on death
* foxfire dissipating / destruction on death
* projectiles that explode on death and slowly dissipate
* plant withering, briar patch withering
* blocks of ice / pillars of salt self-destruction

Misc related stuff:
* correct death descriptions for self-destructing monsters
* ballisto immunity to spores

There are probably still some weird monsters this doesn't cover, e.g.
spatial maelstroms have a weird self-banishing behavior that I need to
look at.