feat: custom names for hellbinder/cloud mage
This is normally done in lua with a random crawl name. For now, I have
decided to make these two monsters use the player name. (Undecided if
this applies to pan lords.) This was surprisingly fiddly to get right,
so there may still be issues; because it looked awkward, I have
suppressed the regular name + title screen on the HUD in favor of "Plog
the Cloud Mage" etc.