feat: slightly clearer flavor for pan lords in HUD
Not sure this really works, but it at least provides a clue as to what's
going on. The use of "the" is partly so that the name renders well on
the save browser. It might actually just make more sense to only show
the randname on `A`?

Still of two minds over whether these should use player names or
randnames -- randnames are much more clearly part of the pan lord's
flavor than for cloud mage etc. (For offline, one idea might be on
option to have the pan lord set the *player*'s name, rather than vice
versa.)