feat: animated objects, initial implementation
These turned out to be hugely painful to implement. A monster not in the
level can't hold items, so my first plan of just using dancing weapons
(etc) as-is wasn't viable (it would have been even *more* painful to change
this, or keep the monster instance in env). What I ended up doing was
keeping the defining object in the player's inventory, with them unable
to unwield or remove it. This has the interesting consequence that they
can enchant (and brand) themselves. Because starts are so extremely
random for these, and "dancing weapon" in particular uses a limited
repertoire, I allowed the player to give an itemspec. This also makes it
possible to chose unrands to play as, as long as they are weapons or
body armour.

There's probably a bunch of flavor that's still wrong for these. Also,
balance is a bit all over the place in ways that may be worth smoothing.