feat: monster species dual wielding
This is one of those cases where a whole over-elaborate implementation
is needed to support something that only three monsters can do, but so
be it. This implementation is still pretty basic in some ways, and there
were a lot of corner cases to spot. One issue is that ' swapping doesn't
fully work to return to a dual-wielding setup. Some quirks are designed
to exactly match current monster behavior -- there is no handedness
check, for example, and gyre and gimble acts like a single weapon.

For attack delay, I used the average of the two. The obvious alternative
would be the max of the two, I'm not sure which is a better choice. (If
dual wielders are "too strong", this may be a useful nerf to apply.)

This adds an equip slot and so it (badly) breaks save compat.